+++ src/Microsoft.Identity.Web.DownstreamApi/DownstreamApi.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Diagnostics.CodeAnalysis;
+using System.IO;
 using System.Linq;
 using System.Net.Http;
 using System.Runtime.CompilerServices;
@@ -219,28 +219,32 @@ namespace Microsoft.Identity.Web
             return clonedOptions;
         }
 
-        private static HttpContent? SerializeInput<TInput>(TInput input, DownstreamApiOptions effectiveOptions)
+        internal static HttpContent? SerializeInput<TInput>(TInput input, DownstreamApiOptions effectiveOptions)
         {
-            HttpContent? effectiveInput;
-            if (input is HttpContent)
+            HttpContent? httpContent;
+
+            if (effectiveOptions.Serializer != null)
             {
-                effectiveInput = input as HttpContent;
-            }
-            else if (effectiveOptions.Serializer != null)
-            {
-                effectiveInput = effectiveOptions.Serializer(input);
+                httpContent = effectiveOptions.Serializer(input);
             }
             else
             {
-#pragma warning disable CA2000 // Dispose objects before losing scope
-                effectiveInput = new StringContent(JsonSerializer.Serialize(input), Encoding.UTF8, "application/json");
-#pragma warning restore CA2000 // Dispose objects before losing scope
+                // if the input is already an HttpContent, it's used as is, and should already contain a ContentType.
+                httpContent = input switch
+                {
+                    HttpContent content => content,
+                    string str when !string.IsNullOrEmpty(effectiveOptions.ContentType) && effectiveOptions.ContentType.StartsWith("text", StringComparison.OrdinalIgnoreCase) => new StringContent(str),
+                    string str => new StringContent(JsonSerializer.Serialize(str), Encoding.UTF8, "application/json"),
+                    byte[] bytes => new ByteArrayContent(bytes),
+                    Stream stream => new StreamContent(stream),
+                    null => null,
+                    _ => new StringContent(JsonSerializer.Serialize(input), Encoding.UTF8, "application/json"),
+                };
             }
-
-            return effectiveInput;
+            return httpContent;
         }
 
-        private static async Task<TOutput?> DeserializeOutput<TOutput>(HttpResponseMessage response, DownstreamApiOptions effectiveOptions)
+        internal static async Task<TOutput?> DeserializeOutput<TOutput>(HttpResponseMessage response, DownstreamApiOptions effectiveOptions)
              where TOutput : class
         {
             try
@@ -265,18 +269,33 @@ namespace Microsoft.Identity.Web
                 return default;
             }
 
+            string? mediaType = content.Headers.ContentType?.MediaType;
+
             if (effectiveOptions.Deserializer != null)
             {
                 return effectiveOptions.Deserializer(content) as TOutput;
             }
+            else if (typeof(TOutput).IsAssignableFrom(typeof(HttpContent)))
+            {
+                return content as TOutput;
+            }
             else
             {
                 string stringContent = await content.ReadAsStringAsync();
-                return JsonSerializer.Deserialize<TOutput>(stringContent, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+                if (mediaType == "application/json")
+                {
+                    return JsonSerializer.Deserialize<TOutput>(stringContent, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+                }
+                if (mediaType != null && !mediaType.StartsWith("text/", StringComparison.OrdinalIgnoreCase))
+                {
+                    // Handle other content types here
+                    throw new NotSupportedException("Content type not supported. Provide your own deserializer. ");
+                }
+                return stringContent as TOutput;
             }
         }
 
-        private async Task<HttpResponseMessage> CallApiInternalAsync(
+        internal async Task<HttpResponseMessage> CallApiInternalAsync(
             string? serviceName,
             DownstreamApiOptions effectiveOptions,
             bool appToken,
@@ -287,14 +306,14 @@ namespace Microsoft.Identity.Web
             // Downstream API URI
             string apiUrl = effectiveOptions.GetApiUrl();
 
-            using HttpClient client = string.IsNullOrEmpty(serviceName) ? _httpClientFactory.CreateClient() : _httpClientFactory.CreateClient(serviceName);
-
             // Create an HTTP request message
             using HttpRequestMessage httpRequestMessage = new(
                 new HttpMethod(effectiveOptions.HttpMethod),
                 apiUrl);
 
             await UpdateRequestAsync(httpRequestMessage, content, effectiveOptions, appToken, user, cancellationToken);
+
+            using HttpClient client = string.IsNullOrEmpty(serviceName) ? _httpClientFactory.CreateClient() : _httpClientFactory.CreateClient(serviceName);
 
             // Send the HTTP message           
             var downstreamApiResult = await client.SendAsync(httpRequestMessage, cancellationToken).ConfigureAwait(false);
@@ -319,7 +338,7 @@ namespace Microsoft.Identity.Web
             return downstreamApiResult;
         }
 
-        private async Task UpdateRequestAsync(
+        internal async Task UpdateRequestAsync(
             HttpRequestMessage httpRequestMessage,
             HttpContent? content,
             DownstreamApiOptions effectiveOptions,
@@ -351,6 +370,10 @@ namespace Microsoft.Identity.Web
             else
             {
                 Logger.UnauthenticatedApiCall(_logger, null);
+            }
+            if (!string.IsNullOrEmpty(effectiveOptions.AcceptHeader))
+            {
+                httpRequestMessage.Headers.Accept.ParseAdd(effectiveOptions.AcceptHeader);
             }
             // Opportunity to change the request message
             effectiveOptions.CustomizeHttpRequestMessage?.Invoke(httpRequestMessage);